refactor(files): Migrate from event listener to class based declarative settings
